requirements - pipreqs.txt
mindspore==2.0.0rc1
numpy==1.15.1
numpy==1.20.3
numpy==1.21.6
numpy==1.23.2
numpy==1.24.2
numpy==1.24.3
numpy==1.21.5
opencv_python==4.7.0.72
Pillow==10.0.1
Pillow==8.3.1
Pillow==9.2.0
Pillow==9.3.0
Pillow==9.4.0
Pillow==9.5.0
Pillow==9.0.1
Pillow==11.0.0
scipy==1.9.3
scipy==1.7.3
tensorboardX==2.6
tensorboardX==2.6.2.2
torch==1.12.0
torch==1.9.0
torch==2.3.0
torch==2.5.0
torch==1.11.0
torchvision==0.10.0
torchvision==0.13.0
torchvision==0.18.1a0
torchvision==0.12.0
